refactor(dashboard): extract board surface modules
