+++ dashboard/src/components/board/index.ts
@@ -1,0 +1,8 @@
+export { BoardSurface } from './board-surface'
+export * from './board-state'
+export {
+  CommentThread,
+  PostDetail,
+  countCommentDescendants,
+  filterCommentTree,
+} from './post-detail'
